requirements.txt
# Web App & UI
streamlit
python-dotenv

# AI & RAG
openai
langchain-text-splitters
sentence-transformers
faiss-cpu
torch
transformers
Pillow
open-clip-torch

# Data Processing & Analytics
pandas
numpy
scikit-learn
nltk
matplotlib
seaborn
tqdm

# Scraping
requests
beautifulsoup4
newspaper3k
lxml_html_clean